Move some functionality into default_input_representation.py
