+++ DeepCrazyhouse/src/domain/variants/crazyhouse/v2/input_representation.py
@@ -8,10 +8,9 @@ Input representation 2.0 for crazyhouse based on the input representation 1.0 of
 """
 
 import chess
-from DeepCrazyhouse.src.domain.variants.classical_chess.input_representation import board_to_planes as\
-    board_to_planes_chess_v1
+from DeepCrazyhouse.src.domain.variants.default_input_representation import default_board_to_planes
 from DeepCrazyhouse.src.domain.variants.constants import MODE_CRAZYHOUSE
-from DeepCrazyhouse.src.domain.variants.input_representation import _set_crazyhouse_info
+from DeepCrazyhouse.src.domain.variants.default_input_representation import _set_crazyhouse_info
 
 
 def board_to_planes(board: chess.Board, board_occ, normalize=True, last_moves=None):
@@ -48,6 +47,6 @@ def board_to_planes(board: chess.Board, board_occ, normalize=True, last_moves=No
 
     """
 
-    planes = board_to_planes_chess_v1(board, board_occ, normalize, last_moves)
+    planes = default_board_to_planes(board, board_occ, normalize, last_moves)
     _set_crazyhouse_info(board, MODE_CRAZYHOUSE, planes, normalize)
     return planes
